Renamed Motd to MotD
